refactor(docs): #704 PR-E — extract _detail submodules (FINAL)
Move the draft-detail page handlers + helpers + modal scripts +
versioning UI out of `app/docs/routes/__init__.py` into three
focused submodules. After this PR, `__init__.py` is a 272-line
registration shim that just re-exports submodule symbols and
mounts the routes — DoD met for #704.

Module layout (post-PR-E):

    _shared.py          — cross-cutting helpers (PR-B)
    _status_tracker.py  — pipeline-stage tracker (PR-B)
    _upload.py          — GET /drafts/new + POST /drafts (PR-C)
    _list.py            — GET /drafts + filter machinery (PR-D)
    _lifecycle.py       — POST /delete + /keep (#703 PR-A)
    _detail.py          — GET /drafts/{id} + status/actions/link-vtk (PR-E)
    _detail_modals.py   — delete + link-vtk modal scripts (PR-E)
    _detail_versions.py — version timeline + diff page (PR-E)

`_detail.py` imports `_validate_parent_vtk_fk` from `_upload.py`
(cross-submodule import — fine, avoids the package-level circular).
`_detail.py` imports `_version_timeline_rows` /
`_version_timeline_section` from `_detail_versions.py` so the
detail page can compose the timeline section without pulling in
the diff page's heavier dependency graph.

Test patch updates: 13 test classes touched. Package-level patches
(`patch("app.docs.routes.fetch_draft")`) now go through the
appropriate submodule (`patch("app.docs.routes._detail.fetch_draft")`
for detail/status/actions/link-vtk; `_detail_versions.fetch_draft`
for diff-page tests). Patch-path contract pinned in
`tests/test_docs_routes_patch_paths.py` (10 → 13 tests).

Final line counts:
    __init__.py           272
    _detail.py            843
    _detail_modals.py     164
    _detail_versions.py   433
    _lifecycle.py         229
    _list.py              688
    _shared.py            243
    _status_tracker.py    305
    _upload.py            698
    total                3875

Verification:
    uv run pytest tests/test_docs_routes_registered.py — 1 passed
    uv run pytest tests/test_docs_routes_patch_paths.py — 13 passed
    uv run pytest tests/ — 2191 passed, 23 skipped
    uv run ruff check + format — clean
    uv run pyright app/docs/routes/ — 0 errors

Closes #704.
